requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
aiohttp
pymongo
motor
beanie
redis
pymilvus
kafka-python
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langchain-milvus==0.1.6
langchain-core>=0.3.28,<1.0.0  # 兼容 langchain-mcp-adapters
langgraph>=0.2.0 <0.3.0 # LangGraph 状态图
ollama
sentence-transformers
transformers
torch
safetensors
accelerate
FlagEmbedding
PyMuPDF
pypdf2
python-docx
docx2txt
python-multipart
python-pptx
pandas
openpyxl
beautifulsoup4
lxml
striprtf
ebooklib
pytesseract
Pillow
python-dotenv
requests
aiofiles
numpy
umap-learn
scikit-learn
bcrypt
PyJWT
toml
loguru
psutil
gputil
mcp>=1.0.0
# langchain_mcp_adapters  # 暂时注释，版本冲突